+++ src/gods/demeter_reload.lua
@@ -11,26 +11,31 @@ function mod.FrostbiteDamage (victim, functionArgs)
 	})
 	local damageAmount = functionArgs.FrostbiteBaseDmg
 	local frostbiteProjectile = {
-		Name = "DemeterAmmoWind",
+		Name = "DemeterFrostbite",
 		Id = CurrentRun.Hero.ObjectId,
 		DestinationId = victim.ObjectId,
 		FireFromTarget = true,
-		DamageMultiplier = damageAmount / 30, -- divided by projectile base damage
+		DamageMultiplier = damageAmount, -- divided by projectile base damage
 		DataProperties =
 		{
-			Fuse = 0,
-			Range = 200,
-			DamageRadius = 250,
-			DamageRadiusScaleY = 0,
-			DamageRadiusScaleX = 0,
+			DamageRadius = 10,
+			DamageRadiusScaleX = 1.175,
+			DamageRadiusScaleY = 0.56,
+			ImpactVelocity = 0,
+			AffectsFriends = false,
+			AffectsSelf = false,
+			AffectsEnemies = true,
+			MaxVictimZ = 9999,
+			CanHitWithoutDamage = false,
 		},
 	}
 
 	wait( functionArgs.Interval, RoomThreadName )
 	local count = 1
 	while victim and victim.RootActive and not victim.IsDead and not CurrentRun.Hero.IsDead do
-		frostbiteProjectile.DamageMultiplier = (functionArgs.FrostbiteBaseDmg * count) / 30 -- divided by projectile base damage
+		frostbiteProjectile.DamageMultiplier = (functionArgs.FrostbiteBaseDmg * count) -- divided by projectile base damage
 		CreateProjectileFromUnit(frostbiteProjectile)
+		CreateAnimation({ Name = "RadialNovaDemeter", DestinationId = victim.ObjectId })
 		count = count + 1
 		wait( functionArgs.Interval, RoomThreadName )
 	end